requirements.txt
google-adk
fastapi
uvicorn
pydantic
python-dotenv
a2a-sdk
gradio